회원가입한 아이디와 롤 아이디 연동 기능 추가, 서비스 기능 코드들 왜 썻는지 주석 달기
