+++ src/main/java/org/example/hamlol/dto/UserDTO.java
@@ -15,7 +15,7 @@ public class UserDTO {
     public UserEntity toEntity() {
         UserEntity userEntity = new UserEntity();
         // userId는 자동 생성되므로 설정하지 않음
-        userEntity.setUsername(this.username);  // DTO에서 username을 가져와 Entity에 설정
+        userEntity.setUserName(this.username);  // DTO에서 username을 가져와 Entity에 설정
         userEntity.setPassword(this.password);  // DTO에서 password를 가져와 Entity에 설정
         return userEntity;
     }
