+++ src/main/java/org/example/hamlol/entity/UserEntity.java
@@ -14,7 +14,7 @@ public class UserEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long userId;
 
-    private String username;
+    private String userName;
     private String password;
