+++ src/main/java/org/example/hamlol/repository/AccountRepository.java
@@ -4,9 +4,9 @@ import org.example.hamlol.entity.AccountEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
 public interface AccountRepository extends JpaRepository<AccountEntity, String> {
-
 
 
 }
