+++ src/main/java/org/example/hamlol/repository/UserRepository.java
@@ -1,10 +1,18 @@
 package org.example.hamlol.repository;
 
+import org.apache.catalina.User;
 import org.example.hamlol.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
 public interface UserRepository extends JpaRepository<UserEntity, Long> {
+
+    UserEntity findByUserName(String userName);
+
+
+
 
 }
