+++ src/main/java/org/example/hamlol/service/impl/UserServiceImpl.java
@@ -1,4 +1,4 @@
-package org.example.hamlol.serviceimpl;
+package org.example.hamlol.service.impl;
 
 import jakarta.transaction.Transactional;
 import org.example.hamlol.dto.UserDTO;
@@ -11,6 +11,7 @@ import org.springframework.stereotype.Service;
 @Service
 public class UserServiceImpl implements UserService {
 
+    // UserRepository를 final 필드로 선언하여 의존성을 주입빋아 DB작업에 사용
     private final UserRepository userRepository;
 
     @Autowired
@@ -20,7 +21,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional
-    public void addUser(UserDTO userDTO) {
+    public void addUser(UserDTO userDTO) { // 클라이언트로부터 받은 UserDTO를 기반으로 새로운 정보를 데이터 베이스에 저장
         // DTO -> Entity 변환
         UserEntity userEntity = userDTO.toEntity();
